# [DO NOT MERGE] [pdata] Try append for slice.CopyTo