Trying consolidating the CopyTo logic for slices as suggested in https://github.com/open-telemetry/opentelemetry-collector/pull/13267#discussion_r2168540752

@bogdandrutu looks like the compiler is not that smart and we cannot use the append approach for the complex slices. See the results:

before
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/collector/pdata/internal
cpu: Apple M1 Max
BenchmarkCopyOrigSlice
BenchmarkCopyOrigSlice/0_to_7-10         	 6696460	       187.7 ns/op	     252 B/op	      10 allocs/op
BenchmarkCopyOrigSlice/1_to_7-10         	 6324669	       181.4 ns/op	     252 B/op	      10 allocs/op
BenchmarkCopyOrigSlice/7_to_7-10         	 6145800	       191.6 ns/op	     252 B/op	      10 allocs/op
BenchmarkCopyOrigSlice/10_to_7-10        	 4897490	       247.2 ns/op	     412 B/op	      11 allocs/op
BenchmarkCopyOrigSlice/20_to_7-10        	 3922288	       306.8 ns/op	     572 B/op	      11 allocs/op
BenchmarkCopyOrigSlice/50_to_7-10        	 2775319	       430.0 ns/op	    1148 B/op	      11 allocs/op
PASS
```

after
```
BenchmarkCopyOrigSlice/0_to_7-10         	 6255495	       188.3 ns/op	     252 B/op	      10 allocs/op
BenchmarkCopyOrigSlice/1_to_7-10         	 5359400	       186.6 ns/op	     252 B/op	      10 allocs/op
BenchmarkCopyOrigSlice/7_to_7-10         	 5916748	       206.2 ns/op	     252 B/op	      10 allocs/op
BenchmarkCopyOrigSlice/10_to_7-10        	 4371021	       264.9 ns/op	     476 B/op	      11 allocs/op
BenchmarkCopyOrigSlice/20_to_7-10        	 3081704	       395.6 ns/op	     924 B/op	      12 allocs/op
BenchmarkCopyOrigSlice/50_to_7-10        	 1881642	       631.2 ns/op	    1948 B/op	      13 allocs/op
PASS
```

10 allocations are spent to regenerate the destination slice